Removed problematic base elements (components, service, module). Now everything is separated as a clear copy. After all, we should try again to use reusable elements. #48
