--- .circleci/config.yml
@@ -1,47 +1,0 @@
-# Javascript Node CircleCI 2.0 configuration file
-#
-# Check https://circleci.com/docs/2.0/language-javascript/ for more details
-#
-version: 2
-jobs:
-  build:
-    docker:
-      - image: angular/ngcontainer:latest
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      # Download and cache dependencies
-      - restore_cache:
-          keys:
-          - v1-dependencies-{{ checksum "yarn.lock" }}
-
-      - run:
-          name: Version Info
-          command: |
-            node --version \
-              && yarn --version \
-              && git --version
-      - run: yarn install
-
-      - save_cache:
-          paths:
-            - node_modules
-          key: v1-dependencies-{{ checksum "yarn.lock" }}
-
-      # Build the libraries
-      - run: yarn build:lib
-      # Build the demo apps (from library packages)
-      - run: yarn build:app
-
-      # Run unit tests!
-      # - run: xvfb-run -a yarn ci:test
-
-      # Run e2e tests!
-      # - run: xvfb-run -a yarn ci:e2e
-
-      # Store build artefacts on circle
-      - store_artifacts:
-          path: ~/repo/dist
-          destination: dist
