+++ lib/ce/src/hot-table.module.ts
@@ -1,14 +1,10 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { HotTableComponent } from './hot-table.component';
 import { HotColumnComponent } from './hot-column.component';
 import { HotTableRegisterer } from './hot-table-registerer.service';
 
 @NgModule({
-  imports: [
-    CommonModule
-  ],
   declarations: [
     HotTableComponent,
     HotColumnComponent,
