+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "angular-handsontable",
-  "version": "1.0.0-beta3",
+  "version": "1.0.0-beta4",
   "description": "Official Angular.io directive for Handsontable",
   "repository": {
     "type": "git",
