+++ lib/pro/src/public_api.ts
@@ -1,4 +1,5 @@
 export * from './hot-table.component';
 export * from './hot-column.component';
 export * from './hot-table-registerer.service';
+export * from './hot-settings-resolver.service';
 export * from './hot-table.module';
