--- renovate.json
@@ -1,38 +1,0 @@
-{
-  "pinVersions": false,
-  "ignoreUnstable": false,
-  "semanticCommits": true,
-  "semanticPrefix": "build:",
-  "commitMessage": "{{semanticPrefix}}update {{depName}} to version {{newVersion}}",
-  "packageFiles": [
-    "package.json",
-    "lib/package.json"
-  ],
-  "packages": [
-    {
-      "packagePattern": "^@angular\/.*",
-      "groupName": "angular",
-      "pinVersions": false
-    },
-    {
-      "packagePattern": "^@types\/.*",
-      "groupName": "type definitions",
-      "pinVersions": false
-    },
-    {
-      "packagePattern": "^jasmine.*",
-      "groupName": "jasmine",
-      "pinVersions": false
-    },
-    {
-      "packagePattern": "^karma.*",
-      "groupName": "karma",
-      "pinVersions": false
-    },
-    {
-      "packagePattern": "^protractor.*",
-      "groupName": "protractor",
-      "pinVersions": false
-    }
-  ]
-}
