+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  template: ``
+  template: `<hot-table></hot-table>`
 })
 export class AppComponent {
